Modified tenant property retrieval logic to work according to current context
