+++ src/main/java/org/oscm/identity/oidc/tenant/PropertyConfigurationPolicy.java
@@ -1,37 +1,34 @@
+/**
+ * *****************************************************************************
+ *
+ * <p>Copyright FUJITSU LIMITED 2019
+ *
+ * <p>Creation Date: June 13, 2019
+ *
+ * <p>*****************************************************************************
+ */
 package org.oscm.identity.oidc.tenant;
 
 import lombok.extern.slf4j.Slf4j;
-import org.oscm.identity.error.TenantConfigurationException;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import java.io.*;
 import java.util.Properties;
 
 @Component
 @Slf4j
 public class PropertyConfigurationPolicy implements TenantConfigurationPolicy {
 
-  @Value("${tenant.config.directory.relative.path}")
-  private String tenantConfigDirectoryRelativePath;
+  private TenantPropertyFileRetriever propertyFileRetriever;
+
+  @Autowired
+  public PropertyConfigurationPolicy(TenantPropertyFileRetriever propertyFileRetriever) {
+    this.propertyFileRetriever = propertyFileRetriever;
+  }
 
   @Override
   public TenantConfiguration loadTenant(String tenantId) {
-
-    Properties properties = new Properties();
-    String propertyFilePath = getPropertiesFilePathForTenant(tenantId);
-
-    try {
-      InputStream inputStream = getFileInputSteam(propertyFilePath);
-      properties.load(inputStream);
-    } catch (IOException exc) {
-      TenantConfigurationException ex =
-          new TenantConfigurationException(
-              "Tenant configuration [tenantId = " + tenantId + "] could not be loaded", exc);
-
-      log.error(ex.getMessage(), exc);
-      throw ex;
-    }
+    Properties properties = propertyFileRetriever.retrievePropertiesFor(tenantId);
 
     TenantConfiguration configuration = new TenantConfiguration();
     configuration.setTenantId(tenantId);
@@ -42,21 +39,5 @@ public class PropertyConfigurationPolicy implements TenantConfigurationPolicy {
     configuration.setOidConfigUrl(properties.getProperty("oidc.openidConfigurationUrl"));
 
     return configuration;
-  }
-
-  private String getPropertiesFilePathForTenant(String tenantId) {
-    return getTenantConfigDirectoryPath() + "tenant-" + tenantId + ".properties";
-  }
-
-  private String getTenantConfigDirectoryPath() {
-    File classPath = new File(System.getProperty("java.class.path"));
-    String jarAbsolutePath = classPath.getAbsolutePath();
-
-    return jarAbsolutePath.substring(0, jarAbsolutePath.length() - classPath.getName().length())
-        + tenantConfigDirectoryRelativePath;
-  }
-
-  protected InputStream getFileInputSteam(String propertyFilePath) throws FileNotFoundException {
-    return new FileInputStream(propertyFilePath);
   }
 }
