+++ src/main/resources/tenant-default.properties
@@ -1,0 +1,5 @@
+oidc.provider=default
+oidc.clientId=defaultClientId
+oidc.authUrl=defaultAuthUrl
+oidc.idTokenRedirectUrl=defaultIdTokenRedirectUrl
+oidc.openidConfigurationUrl=defaultOIDConfigUrl
