+++ src/test/resources/application.properties
@@ -1,3 +1,5 @@
+spring.profiles.active=test
+
 tenant.config.directory.relative.path=/config/tenants/
 
 auth.signing.algorithm.type=RS256
